+++ src/solver_n_dimensional/solver/rootfinder.rs
@@ -2,17 +2,19 @@ use std::fmt;
 
 extern crate nalgebra;
 
+use crate::errors;
 use crate::iteratives;
 use crate::iteratives::Iterative;
 use crate::model;
+use crate::model::ModelError;
 use crate::residuals;
 
-use super::greenstadt_second_method_udpate_jac;
-use super::{broyden_first_method_udpate_inv_jac, broyden_second_method_udpate_inv_jac};
-use super::{broyden_first_method_udpate_jac, broyden_second_method_udpate_jac};
-use super::{jacobian_evaluation, JacobianMatrix, SolverParameters};
-use super::{quasi_method_update_inv_jac, quasi_method_update_jac};
-use super::{QuasiNewtonMethod, ResolutionMethod, UpdateQuasiNewtonMethod};
+use super::{
+    approximate_inv_jacobian, approximate_jacobian, evaluate_jacobian_from_analytical_function,
+    evaluate_jacobian_from_finite_difference, JacobianMatrix, SolverParameters,
+};
+
+use super::{QuasiNewtonMethod, ResolutionMethod};
 
 /// Solver for rootfinding
 ///
@@ -41,11 +43,10 @@ where
     iter: usize,
     solver_log: Option<super::log::SolverLog>,
     jacobian: JacobianMatrix<D>,
-    compute_jac_next_iter: bool,
-    last_iter_with_computed_jacobian: usize,
     iteratives_step_size: Option<nalgebra::OVector<f64, D>>,
     residuals_step_size: Option<nalgebra::OVector<f64, D>>,
     residuals_values_current: Option<nalgebra::OVector<f64, D>>,
+    valid_last_model_evaluation: bool,
 }
 
 impl<'a, T, D> RootFinder<'a, T, D>
@@ -90,11 +91,10 @@ where
         }
 
         let jacobian = JacobianMatrix::new();
-        let compute_jac_next_iter = true;
-        let last_iter_with_computed_jacobian = 0;
         let iteratives_step_size = None;
         let residuals_step_size = None;
         let residuals_values_current = None;
+        let valid_last_model_evaluation = true;
 
         RootFinder {
             parameters,
@@ -105,11 +105,10 @@ where
             iter,
             solver_log,
             jacobian,
-            compute_jac_next_iter,
-            last_iter_with_computed_jacobian,
             iteratives_step_size,
             residuals_step_size,
             residuals_values_current,
+            valid_last_model_evaluation,
         }
     }
 
@@ -159,145 +158,103 @@ where
             .evaluate_stopping_residuals(&residuals_values)
     }
 
-    fn compute_jac_func<M>(&mut self, model: &mut M)
+    fn compute_jac<M>(&mut self, model: &mut M) -> Result<(), errors::SolverInternalError<M, D>>
     where
         M: model::Model<D>,
     {
-        let residuals_values = model.get_residuals();
-
-        let jacobians = model.get_jacobian();
-        let normalization_method = self.residuals_config.get_update_methods();
-        self.jacobian
-            .update_jacobian(jacobians.normalize(&residuals_values, &normalization_method));
-    }
-
-    fn compute_jac_fd<M>(&mut self, model: &mut M)
-    where
-        M: model::Model<D>,
-    {
-        let iters_values = model.get_iteratives();
-
-        let perturbations = self.iters_params.compute_perturbations(&iters_values);
-
-        self.jacobian.update_jacobian(jacobian_evaluation(
-            model,
-            &perturbations,
-            &(self.residuals_config),
-        ));
-    }
-
-    fn compute_jac<M>(&mut self, model: &mut M)
-    where
-        M: model::Model<D>,
-    {
-        if model.jacobian_provided() {
-            self.compute_jac_func(model);
+        let successful_jac_computation = if model.jacobian_provided() {
+            evaluate_jacobian_from_analytical_function(
+                &mut self.jacobian,
+                model,
+                self.residuals_config,
+            )
         } else {
-            self.compute_jac_fd(model);
-        }
+            evaluate_jacobian_from_finite_difference(
+                &mut self.jacobian,
+                model,
+                self.iters_params,
+                self.residuals_config,
+            )
+        };
 
-        self.compute_jac_next_iter = false;
-        self.last_iter_with_computed_jacobian = self.iter;
+        match successful_jac_computation {
+            Ok(())
+            | Err(errors::SolverInternalError::InvalidJacobianError(
+                ModelError::InaccurateValuesError(_),
+            )) => Ok(()),
+            Err(model_error) => Err(model_error),
+        }
     }
 
-    fn compute_newton_raphson_step<M>(&mut self, model: &mut M) -> nalgebra::OVector<f64, D>
+    fn compute_newton_raphson_step<M>(
+        &mut self,
+        model: &mut M,
+    ) -> Result<nalgebra::OVector<f64, D>, crate::errors::SolverInternalError<M, D>>
     where
         M: model::Model<D>,
     {
-        self.compute_jac(model);
+        let successful_jac_computation = self.compute_jac(model);
 
         if self.debug {
             self.jac_to_log();
         }
-
-        self.compute_next_from_inv_jac(model)
+        match successful_jac_computation {
+            Ok(()) => Ok(self.compute_next_from_inv_jac(model)),
+            Err(error) => Err(error),
+        }
     }
 
-    fn approximate_jacobian(&mut self, method: UpdateQuasiNewtonMethod) {
-        let jac_next = match method {
-            UpdateQuasiNewtonMethod::BroydenFirstMethod => broyden_first_method_udpate_jac(
-                self.jacobian.get_jacobian().as_ref().unwrap(),
-                self.iteratives_step_size.as_ref().unwrap(),
-                self.residuals_step_size.as_ref().unwrap(),
-            ),
-            UpdateQuasiNewtonMethod::BroydenSecondMethod => broyden_second_method_udpate_jac(
-                self.jacobian.get_jacobian().as_ref().unwrap(),
-                self.iteratives_step_size.as_ref().unwrap(),
-                self.residuals_step_size.as_ref().unwrap(),
-            ),
-            UpdateQuasiNewtonMethod::GreenstadtFirstMethod => quasi_method_update_jac(
-                self.jacobian.get_jacobian().as_ref().unwrap(),
-                self.iteratives_step_size.as_ref().unwrap(),
-                self.residuals_step_size.as_ref().unwrap(),
-                self.residuals_values_current.as_ref().unwrap(),
-            ),
-            UpdateQuasiNewtonMethod::GreenstadtSecondMethod => {
-                let c = self.jacobian.get_inverse().as_ref().unwrap()
-                    * self.residuals_step_size.as_ref().unwrap();
-                greenstadt_second_method_udpate_jac(
-                    self.jacobian.get_jacobian().as_ref().unwrap(),
-                    self.iteratives_step_size.as_ref().unwrap(),
-                    self.residuals_step_size.as_ref().unwrap(),
-                    &c,
-                )
-            }
-        };
-
-        self.jacobian.update_jacobian(jac_next);
-    }
-
-    fn approximate_inv_jacobian(&mut self, method: UpdateQuasiNewtonMethod) {
-        let inv_jac_next = match method {
-            UpdateQuasiNewtonMethod::BroydenFirstMethod => broyden_first_method_udpate_inv_jac(
-                self.jacobian.get_inverse().as_ref().unwrap(),
-                self.iteratives_step_size.as_ref().unwrap(),
-                self.residuals_step_size.as_ref().unwrap(),
-            ),
-            UpdateQuasiNewtonMethod::BroydenSecondMethod => broyden_second_method_udpate_inv_jac(
-                self.jacobian.get_inverse().as_ref().unwrap(),
-                self.iteratives_step_size.as_ref().unwrap(),
-                self.residuals_step_size.as_ref().unwrap(),
-            ),
-            UpdateQuasiNewtonMethod::GreenstadtFirstMethod => quasi_method_update_inv_jac(
-                self.jacobian.get_inverse().as_ref().unwrap(),
-                self.iteratives_step_size.as_ref().unwrap(),
-                self.residuals_step_size.as_ref().unwrap(),
-                self.residuals_values_current.as_ref().unwrap(),
-            ),
-            UpdateQuasiNewtonMethod::GreenstadtSecondMethod => {
-                let c = self.jacobian.get_inverse().as_ref().unwrap().transpose()
-                    * self.jacobian.get_inverse().as_ref().unwrap()
-                    * self.residuals_step_size.as_ref().unwrap();
-                quasi_method_update_inv_jac(
-                    self.jacobian.get_inverse().as_ref().unwrap(),
-                    self.iteratives_step_size.as_ref().unwrap(),
-                    self.residuals_step_size.as_ref().unwrap(),
-                    &c,
-                )
-            }
-        };
-
-        self.jacobian.update_inverse(inv_jac_next);
-    }
-
-    fn compute_quasi_newton_step<M>(
+    /// Perform the jacobian evaluation
+    ///
+    /// Based on the resolution method:
+    /// - the jacobian can be computed and inverted
+    /// - the jacobian can be approximated and inverted
+    /// - the inverse of the jacobian can be approximated
+    fn evaluate_jacobian_quasi_newton_step<M>(
         &mut self,
         model: &mut M,
         resolution_method: QuasiNewtonMethod,
-    ) -> nalgebra::OVector<f64, D>
+    ) -> Result<(), crate::errors::SolverInternalError<M, D>>
     where
         M: model::Model<D>,
     {
-        if self.compute_jac_next_iter {
-            self.compute_jac(model);
+        if self.jacobian.compute_jacobian() {
+            let successful_jac_computation = self.compute_jac(model);
+
+            match successful_jac_computation {
+                Ok(()) => (),
+                Err(error) => {
+                    if self.debug {
+                        self.jac_to_log();
+                    }
+                    return Err(error);
+                }
+            }
         } else {
             match resolution_method {
                 QuasiNewtonMethod::StationaryNewton => (),
                 QuasiNewtonMethod::JacobianUpdate(method) => {
-                    self.approximate_jacobian(method);
+                    match approximate_jacobian(
+                        &mut self.jacobian,
+                        method,
+                        self.iteratives_step_size.as_ref().unwrap(),
+                        self.residuals_step_size.as_ref().unwrap(),
+                        self.residuals_values_current.as_ref().unwrap(),
+                    ) {
+                        Ok(()) => (),
+                        Err(_) => {
+                            return Err(errors::SolverInternalError::InvalidJacobianInverseError)
+                        }
+                    }
                 }
                 QuasiNewtonMethod::InverseJacobianUpdate(method) => {
-                    self.approximate_inv_jacobian(method);
+                    approximate_inv_jacobian(
+                        &mut self.jacobian,
+                        method,
+                        self.iteratives_step_size.as_ref().unwrap(),
+                        self.residuals_step_size.as_ref().unwrap(),
+                        self.residuals_values_current.as_ref().unwrap(),
+                    );
                 }
             };
         }
@@ -306,7 +263,26 @@ where
             self.jac_to_log();
         }
 
-        self.compute_next_from_inv_jac(model)
+        Ok(())
+    }
+
+    fn compute_quasi_newton_step<M>(
+        &mut self,
+        model: &mut M,
+        resolution_method: QuasiNewtonMethod,
+    ) -> Result<nalgebra::OVector<f64, D>, crate::errors::SolverInternalError<M, D>>
+    where
+        M: model::Model<D>,
+    {
+        match self.evaluate_jacobian_quasi_newton_step(model, resolution_method) {
+            Ok(()) => Ok(self.compute_next_from_inv_jac(model)),
+            Err(crate::errors::SolverInternalError::InvalidJacobianError(error)) => Err(
+                crate::errors::SolverInternalError::InvalidJacobianError(error),
+            ),
+            Err(crate::errors::SolverInternalError::InvalidJacobianInverseError) => {
+                Err(crate::errors::SolverInternalError::InvalidJacobianInverseError)
+            }
+        }
     }
 
     fn compute_next_from_inv_jac<M>(&self, model: &M) -> nalgebra::OVector<f64, D>
@@ -338,9 +314,9 @@ where
         if max_error_next > max_error {
             // see documentation of the `SolverParameters` struct
             if self.parameters.get_resolution_method() != ResolutionMethod::NewtonRaphson
-                && self.last_iter_with_computed_jacobian != self.iter
+                && self.jacobian.is_jacobian_approximated()
             {
-                self.compute_jac_next_iter = true;
+                self.jacobian.force_jacobian_computation();
                 if self.debug {
                     self.recompute_jacobian_to_log();
                 }
@@ -349,11 +325,11 @@ where
                 let damped_guess =
                     current_guess * (1.0 - damping_factor) + proposed_guess * damping_factor;
                 model.set_iteratives(&damped_guess);
-                model.evaluate();
+                model.evaluate().unwrap();
                 *errors_next = self.evaluate_errors(model);
 
                 if self.debug {
-                    self.damping_to_log(model, &errors_next);
+                    self.damping_to_log(model, errors_next);
                 }
             }
         }
@@ -363,7 +339,7 @@ where
         &mut self,
         model: &mut M,
         proposed_guess: &nalgebra::OVector<f64, D>,
-    ) -> nalgebra::OVector<f64, D>
+    ) -> Result<nalgebra::OVector<f64, D>, errors::SolverError<M, D>>
     where
         M: model::Model<D>,
     {
@@ -372,7 +348,18 @@ where
         let current_guess = model.get_iteratives();
 
         model.set_iteratives(proposed_guess);
-        model.evaluate();
+        match model.evaluate() {
+            Ok(()) => {
+                self.valid_last_model_evaluation = true;
+            }
+            Err(ModelError::InaccurateValuesError(_)) => {
+                self.valid_last_model_evaluation = false;
+            }
+            Err(e) => {
+                self.valid_last_model_evaluation = false;
+                return Err(errors::SolverError::ModelEvaluationError(e));
+            }
+        }
         let mut errors_next = self.evaluate_errors(model);
 
         if self.debug {
@@ -384,7 +371,7 @@ where
                 model,
                 max_error,
                 &current_guess,
-                &proposed_guess,
+                proposed_guess,
                 &mut errors_next,
             );
         }
@@ -399,16 +386,32 @@ where
             }
         };
 
-        errors_next
+        Ok(errors_next)
     }
 
     /// The core function performing the resolution on a given `Model`
-    pub fn solve<M>(&mut self, model: &mut M)
+    pub fn solve<M>(&mut self, model: &mut M) -> Result<(), crate::errors::SolverError<M, D>>
     where
         M: model::Model<D>,
     {
         model.set_iteratives(&self.initial_guess);
-        model.evaluate();
+
+        // The first evaluation must yield usuable values
+        // However, then don't need to be accurate
+        match model.evaluate() {
+            Ok(()) => (),
+            Err(ModelError::InaccurateValuesError(_)) => (),
+            Err(ModelError::UnusableValuesError(error)) => {
+                return Err(crate::errors::SolverError::ModelInitialEvaluationError(
+                    error.to_string(),
+                ))
+            }
+            Err(ModelError::UnrecoverableError(error)) => {
+                return Err(crate::errors::SolverError::ModelInitialEvaluationError(
+                    error.to_string(),
+                ))
+            }
+        }
 
         let mut errors = self.evaluate_errors(model);
         let mut max_error = errors.amax();
@@ -418,18 +421,8 @@ where
             self.iteration_to_log(model, &errors);
         }
 
-        // Warning: unrolling by hand the first iteration as the following code does
+        // Warning: unrolling by hand the first iteration (which is always a Newton-Raphson step)
         //          is actually slowing down the code (run benchmarks to see it)
-        //
-        // first iteration: always a Newton-Raphson step
-        //    if max_error > self.parameters.get_tolerance() {
-        //        self.iter += 1;
-        //        let proposed_guess = self.compute_newton_raphson_step(model);
-        //        errors = self.update_model(model, &proposed_guess);
-        //        max_error = errors.amax();
-        //    }
-
-        // other iterations
         while max_error > self.parameters.get_tolerance()
             && self.iter < self.parameters.get_max_iter()
         {
@@ -442,8 +435,25 @@ where
                 }
             };
 
-            errors = self.update_model(model, &proposed_guess);
+            match proposed_guess {
+                Ok(value) => match self.update_model(model, &value) {
+                    Ok(value) => errors = value,
+                    Err(e) => return Err(e),
+                },
+                Err(error) => {
+                    return Err(errors::SolverError::JacobianError(error));
+                }
+            }
+
             max_error = errors.amax();
+        }
+
+        if max_error > self.parameters.get_tolerance() {
+            Err(crate::errors::SolverError::NonConvergenceError)
+        } else if self.valid_last_model_evaluation {
+            Ok(())
+        } else {
+            Err(crate::errors::SolverError::FinalEvaluationError)
         }
     }
 
@@ -471,7 +481,7 @@ where
 
     fn recompute_jacobian_to_log(&self) {
         self.solver_log.as_ref().unwrap().add_content(
-            &"Iteration refused, the jacobian will be recomputed at the next iteration\n\n",
+            "Iteration refused, the jacobian will be recomputed at the next iteration\n\n",
         );
     }