+++ tests/jacobian/broyden.rs
@@ -2,7 +2,7 @@ extern crate newton_rootfinder;
 use newton_rootfinder as nrf;
 use nrf::model::Model;
 use nrf::residuals;
-use nrf::solver::jacobian_evaluation;
+use nrf::solver::compute_jacobian_from_finite_difference;
 
 use crate::common::float_matrix_comparison;
 use util::test_cases::broyden1965;
@@ -17,13 +17,14 @@ fn jacobian_evaluation_broyden1965_case5() {
     let mut user_model =
         nrf::model::UserModelFromFunction::new(problem_size, broyden1965::broyden1965_case5);
     user_model.set_iteratives(&inputs);
-    user_model.evaluate();
+    user_model.evaluate().unwrap();
 
     let stopping_residuals = vec![residuals::NormalizationMethod::Abs; problem_size];
     let update_residuals = stopping_residuals.clone();
     let res_config = residuals::ResidualsConfig::new(&stopping_residuals, &update_residuals);
     let perturbations = nalgebra::DVector::from_vec(vec![5e-8; problem_size]);
-    let jac = jacobian_evaluation(&mut user_model, &perturbations, &res_config);
+    let jac = compute_jacobian_from_finite_difference(&mut user_model, &perturbations, &res_config)
+        .unwrap();
     let jac_ref = broyden1965::broyden1965_case5_jac(&inputs);
 
     float_matrix_comparison(&jac, &jac_ref, 1e-6);
@@ -36,13 +37,14 @@ fn jacobian_evaluation_broyden1965_case6() {
     let mut user_model =
         nrf::model::UserModelFromFunction::new(problem_size, broyden1965::broyden1965_case6);
     user_model.set_iteratives(&inputs);
-    user_model.evaluate();
+    user_model.evaluate().unwrap();
 
     let stopping_residuals = vec![residuals::NormalizationMethod::Abs; problem_size];
     let update_residuals = stopping_residuals.clone();
     let res_config = residuals::ResidualsConfig::new(&stopping_residuals, &update_residuals);
     let perturbations = nalgebra::DVector::from_vec(vec![5e-8; problem_size]);
-    let jac = jacobian_evaluation(&mut user_model, &perturbations, &res_config);
+    let jac = compute_jacobian_from_finite_difference(&mut user_model, &perturbations, &res_config)
+        .unwrap();
     let jac_ref = broyden1965::broyden1965_case6_jac(&inputs);
 
     float_matrix_comparison(&jac, &jac_ref, 1e-6);
@@ -59,13 +61,14 @@ fn jacobian_evaluation_broyden1965_case7() {
         nrf::model::UserModelFromFunction::new(problem_size, broyden1965::broyden1965_case7);
 
     user_model.set_iteratives(&inputs);
-    user_model.evaluate();
+    user_model.evaluate().unwrap();
 
     let stopping_residuals = vec![residuals::NormalizationMethod::Abs; problem_size];
     let update_residuals = stopping_residuals.clone();
     let res_config = residuals::ResidualsConfig::new(&stopping_residuals, &update_residuals);
     let perturbations = nalgebra::DVector::from_vec(vec![5e-8; problem_size]);
-    let jac = jacobian_evaluation(&mut user_model, &perturbations, &res_config);
+    let jac = compute_jacobian_from_finite_difference(&mut user_model, &perturbations, &res_config)
+        .unwrap();
     let jac_ref = broyden1965::broyden1965_case7_jac(&inputs);
 
     float_matrix_comparison(&jac, &jac_ref, 1e-6);
@@ -79,13 +82,14 @@ fn jacobian_evaluation_broyden1965_case8() {
         nrf::model::UserModelFromFunction::new(problem_size, broyden1965::broyden1965_case8);
 
     user_model.set_iteratives(&inputs);
-    user_model.evaluate();
+    user_model.evaluate().unwrap();
 
     let stopping_residuals = vec![residuals::NormalizationMethod::Abs; problem_size];
     let update_residuals = stopping_residuals.clone();
     let res_config = residuals::ResidualsConfig::new(&stopping_residuals, &update_residuals);
     let perturbations = nalgebra::DVector::from_vec(vec![5e-8; problem_size]);
-    let jac = jacobian_evaluation(&mut user_model, &perturbations, &res_config);
+    let jac = compute_jacobian_from_finite_difference(&mut user_model, &perturbations, &res_config)
+        .unwrap();
     let jac_ref = broyden1965::broyden1965_case8_jac(&inputs);
 
     float_matrix_comparison(&jac, &jac_ref, 1e-6);
@@ -99,13 +103,14 @@ fn jacobian_evaluation_broyden1965_case9() {
         nrf::model::UserModelFromFunction::new(problem_size, broyden1965::broyden1965_case9);
 
     user_model.set_iteratives(&inputs);
-    user_model.evaluate();
+    user_model.evaluate().unwrap();
 
     let stopping_residuals = vec![residuals::NormalizationMethod::Abs; problem_size];
     let update_residuals = stopping_residuals.clone();
     let res_config = residuals::ResidualsConfig::new(&stopping_residuals, &update_residuals);
     let perturbations = nalgebra::DVector::from_vec(vec![5e-8; problem_size]);
-    let jac = jacobian_evaluation(&mut user_model, &perturbations, &res_config);
+    let jac = compute_jacobian_from_finite_difference(&mut user_model, &perturbations, &res_config)
+        .unwrap();
     let jac_ref = broyden1965::broyden1965_case9_jac(&inputs);
 
     float_matrix_comparison(&jac, &jac_ref, 1e-6);
@@ -119,13 +124,14 @@ fn jacobian_evaluation_broyden1965_case10() {
         nrf::model::UserModelFromFunction::new(problem_size, broyden1965::broyden1965_case10);
 
     user_model.set_iteratives(&inputs);
-    user_model.evaluate();
+    user_model.evaluate().unwrap();
 
     let stopping_residuals = vec![residuals::NormalizationMethod::Abs; problem_size];
     let update_residuals = stopping_residuals.clone();
     let res_config = residuals::ResidualsConfig::new(&stopping_residuals, &update_residuals);
     let perturbations = nalgebra::DVector::from_vec(vec![5e-8; problem_size]);
-    let jac = jacobian_evaluation(&mut user_model, &perturbations, &res_config);
+    let jac = compute_jacobian_from_finite_difference(&mut user_model, &perturbations, &res_config)
+        .unwrap();
     let jac_ref = broyden1965::broyden1965_case10_jac(&inputs);
 
     float_matrix_comparison(&jac, &jac_ref, 1e-6);