+++ tests/user_model/broyden.rs
@@ -20,8 +20,8 @@ fn broyden_case5_jac() {
     );
 
     user_model.set_iteratives(&init_guess);
-    user_model.evaluate();
-    let jacobians = user_model.get_jacobian();
+    user_model.evaluate().unwrap();
+    let jacobians = user_model.get_jacobian().unwrap();
     let (left_jac, right_jac) = jacobians.get_jacobians();
 
     let jac_ref = broyden1965_case5_jac(&init_guess);
@@ -43,8 +43,8 @@ fn broyden_case6_jac() {
     );
 
     user_model.set_iteratives(&init_guess);
-    user_model.evaluate();
-    let jacobians = user_model.get_jacobian();
+    user_model.evaluate().unwrap();
+    let jacobians = user_model.get_jacobian().unwrap();
     let (left_jac, right_jac) = jacobians.get_jacobians();
 
     let jac_ref = broyden1965_case6_jac(&init_guess);
@@ -66,8 +66,8 @@ fn broyden_case7_jac() {
     );
 
     user_model.set_iteratives(&init_guess);
-    user_model.evaluate();
-    let jacobians = user_model.get_jacobian();
+    user_model.evaluate().unwrap();
+    let jacobians = user_model.get_jacobian().unwrap();
     let (left_jac, right_jac) = jacobians.get_jacobians();
 
     let jac_ref = broyden1965_case7_jac(&init_guess);
@@ -89,8 +89,8 @@ fn broyden_case8_jac() {
     );
 
     user_model.set_iteratives(&init_guess);
-    user_model.evaluate();
-    let jacobians = user_model.get_jacobian();
+    user_model.evaluate().unwrap();
+    let jacobians = user_model.get_jacobian().unwrap();
     let (left_jac, right_jac) = jacobians.get_jacobians();
 
     let jac_ref = broyden1965_case8_jac(&init_guess);
@@ -112,8 +112,8 @@ fn broyden_case9_jac() {
     );
 
     user_model.set_iteratives(&init_guess);
-    user_model.evaluate();
-    let jacobians = user_model.get_jacobian();
+    user_model.evaluate().unwrap();
+    let jacobians = user_model.get_jacobian().unwrap();
     let (left_jac, right_jac) = jacobians.get_jacobians();
 
     let jac_ref = broyden1965_case9_jac(&init_guess);
@@ -135,8 +135,8 @@ fn broyden_case10_jac() {
     );
 
     user_model.set_iteratives(&init_guess);
-    user_model.evaluate();
-    let jacobians = user_model.get_jacobian();
+    user_model.evaluate().unwrap();
+    let jacobians = user_model.get_jacobian().unwrap();
     let (left_jac, right_jac) = jacobians.get_jacobians();
 
     let jac_ref = broyden1965_case10_jac(&init_guess);